+++ src/run.rs
@@ -122,7 +122,16 @@ fn make_replaced_out_one(line_offset: usize, caps: &Captures<'_>, fmt: &str) -> 
             }
         } else if b == b'$' {
             // case of "$$"
-            v_out_s.push_str("$");
+            v_out_s.push('$');
+        } else if b == b'{' {
+            // named capture group
+            if let Some(found) = (fmt[(cur + 1)..]).find('}') {
+                let name = &fmt[(cur + 1)..(cur + 1 + found)];
+                if let Some(mat) = caps.name(name) {
+                    v_out_s.push_str(mat.as_str());
+                }
+                cur = cur + 1 + found + 1;
+            }
         } else {
             cur -= 1;
         }