+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 * minimum support version in doc
 * bug: replacement is "$$"
+* bug: named capture groups
 
 ## [0.1.38] (2024-06-20)
 ### Added